abstracted ploting functionality to plot.py for cleaner responsibility.
